Consolidate serialization modules
